Mv ApiWrapper to separate file
